Complete migrating off of JSON data in the doc. Split the parser to animation and parser crates
