+++ crates/dc_squoosh_animation/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "dc_squoosh_animation"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+dc_bundle = { path = "../dc_bundle" }
+thiserror = "1.0"
