+++ support-figma/dc_squoosh_designer/DATA_FORMAT.md
@@ -26,40 +26,37 @@ The root object contains the animation specification for transitioning *to* the 
     "interrupt_type": "None"
   },
   "customKeyframeData": {
-    "TIMELINE_ID": "SERIALIZED_KEYFRAMES_STRING",
-    // This field stores serialized data for custom (user-added) property timelines.
+    "TIMELINE_ID": "{\"targetEasing\":\"Inherit\",\"keyframes\":[{\"fraction\":0.5,\"value\":{\"Scalar\":1},\"easing\":\"Linear\"}]}"
+    // This field stores stringified JSON data for custom (user-added) property timelines.
+    // Figma's API limits plugin data to strings, so the inner object is JSON-serialized again.
     // System-generated timelines (e.g., for 'x', 'y' changes between variants)
     // do not store their keyframe data here, but derive it from variant properties.
-    ...
   }
 }
 ```
 
-To easily and safely persist complex keyframe sequences, `customKeyframeData` stores serialized property timelines strictly as robust, typed JSON objects mirroring the protobuf schema.
+To easily and safely persist complex keyframe sequences, `customKeyframeData` stores serialized property timelines as stringified JSON objects.
 
-### Standard Format (Typed JSON Schema)
+### Structured Timeline Format (Decoded)
 
-Each custom timeline property entry contains a strongly-typed JSON object array representing all keyframes, perfectly aligning with the Rust engine (`CustomTimeline`) and the Protobuf definitions.
+When decoded, the JSON string mapped to `TIMELINE_ID` has the following structure, aligning with the Protobuf definitions:
 
 ```json
 {
-  "customKeyframeData": {
-    "opacity": {
-      "keyframes": [
-        {
-          "fraction": 0.2,
-          "value": { "Scalar": 0.4 },
-          "easing": "Linear"
-        },
-        {
-          "fraction": 0.5,
-          "value": { "Scalar": 0.8 },
-          "easing": "EaseIn"
-        }
-      ]
+  "targetEasing": "Inherit",
+  "keyframes": [
+    {
+      "fraction": 0.2,
+      "value": { "Scalar": 0.4 },
+      "easing": "Linear"
+    },
+    {
+      "fraction": 0.5,
+      "value": { "Scalar": 0.8 },
+      "easing": "EaseIn"
     }
-  }
+  ]
 }
 ```
 
-*Note: All custom timelines enforce this JSON structure natively, yielding superior parsing performance, explicit typing for multi-scalar values (like color or gradient arrays), and complete immunity to legacy string-delimiter collisions.*
+*Note: All custom timelines enforce this JSON structure within the stringified payload, yielding superior parsing performance and explicit typing for multi-scalar values (like color or gradient arrays).*
