+++ support-figma/dc_squoosh_designer/rs/Cargo.toml
@@ -12,6 +12,7 @@ repository.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 log.workspace = true
+dc_squoosh_animation = { path = "../../../crates/dc_squoosh_animation" }
 
 [dev-dependencies]
 env_logger.workspace = true
